Log out the error when reading bower.json fails
Also changed the wording in the error message to be more generic.

After a few minutes of trying to figure out why Preen couldn't find the
bower.json file in the directory I was running it in, I put in the error
logging and saw the problem was actually a a JSON formatting error in my
bower.json file.